Update of Wors Practices page (Spanish version)
Update of page.

Note.-

I noticed that some pages of the documentation in Spanish were badly translated. I don't know if a service translator, like Google, was used but if you read the text and titles of the translated pages you will notice the literal translation.

I don't pretend to be a professional translator. It's just an observation.