Add clearer POM terminology (component objects)
Referring to both an abstraction of the page itself and the components
of that page as page objects can cause confusion, in part because it
requires explaining that you can have more than one page object per
page, and some nuances to that can get lost (as it could be taken to
mean you can have more than one page "page" object per page).

It can also be confusing because an object that represents a component
of the page would be referred to as a "page" object, despite it not
representing a page, but an object that represents the page as a whole
would be referred to as a "page" object, and is representing a page.

Differentiating between them by defining a new term for the objects
that represent components (i.e. page component objects) avoids the
confusion.