Fixed small typo in driver_requirements: code-snipped "import ...firefox..." (#591)[deploy site]
* Fixed small typo in driver_requirements

There was a small typo in the code snipped of the "Firefox" section of "driver_requirements.en.md". Using the capital "F" in the import statement made the Java compiler not find the class "org.openqa.selenium.firefox.FirefoxDriver" and the build fails.

* Fixed small typo in driver_requirements

There was a small typo in the code snipped of the "Firefox" section of "driver_requirements.en.md". Using the capital "F" in the import statement made the Java compiler not find the class "org.openqa.selenium.firefox.FirefoxDriver" and the build fails.

* Fixed small typo in driver_requirements

There was a small typo in the code snipped of the "Firefox" section of "driver_requirements.en.md". Using the capital "F" in the import statement made the Java compiler not find the class "org.openqa.selenium.firefox.FirefoxDriver" and the build fails.

* Fixed small typo in driver_requirements

There was a small typo in the code snipped of the "Firefox" section of "driver_requirements.en.md". Using the capital "F" in the import statement made the Java compiler not find the class "org.openqa.selenium.firefox.FirefoxDriver" and the build fails.

* Fixed small typo in driver_requirements

There was a small typo in the code snipped of the "Firefox" section of "driver_requirements.en.md". Using the capital "F" in the import statement made the Java compiler not find the class "org.openqa.selenium.firefox.FirefoxDriver" and the build fails.

* Fixed small typo in driver_requirements

There was a small typo in the code snipped of the "Firefox" section of "driver_requirements.en.md". Using the capital "F" in the import statement made the Java compiler not find the class "org.openqa.selenium.firefox.FirefoxDriver" and the build fails.

* Fixed small typo in driver_requirements

There was a small typo in the code snipped of the "Firefox" section of "driver_requirements.en.md". Using the capital "F" in the import statement made the Java compiler not find the class "org.openqa.selenium.firefox.FirefoxDriver" and the build fails.

* Fixed small typo in driver_requirements

There was a small typo in the code snipped of the "Firefox" section of "driver_requirements.en.md". Using the capital "F" in the import statement made the Java compiler not find the class "org.openqa.selenium.firefox.FirefoxDriver" and the build fails.

Co-authored-by: JS31096 <12621691+JS31096@users.noreply.github.com>