Update select_elements.ja.md
Replaced the &lt; and &gt; to < and > in a example for python and ruby.
The &lt; and &gt; are not rendered to < and > symbols as they are formatted as code.
It is displayed as '&lt;select&gt;' in the webpage, so made this changes so it will be displayed as '<select>'. 
Please add this proposed changes as it may affect people who are reading examples.