In signInPage there is no Verification going on.
In Line number 191, it is mentioned that SignInPage and HomePage constructors are checking that the expected page is available and ready for requests from the test.
But in the SignInPage example, there is no verification of any sort happening inside the constructor(line no. 97) or anywhere. I have proposed the supposed verification.