Fixed missing await in finders.en.md (#1288)[deploy site]
* Fixed missing await in finders.en.md

findElements method in selenium javascript actually returns a promise which which is supposed to be full-filled in future without this promise being resolved the the other code in the further lines will throw an error. so await should be a compulsion or the code that is supposed to be executed should be put in a then block like shown below.

```
driver.findElements(By.tagName("p")).then(() => {
    // The code that depends on the above element should lie here.
})

```
this can be avoided by simply adding an await as shown. 
```
let element = await driver.findElements(By.tagName("p"))
```

* Update finders.ja.md

* Update finders.pt-br.md

* Update finders.zh-cn.md