Rephrase sentence. Separate two sentences with period.
Replace
```
WebDriver drives a browser natively ..., marks a leaps forward
```
with
```
WebDriver drives a browser natively .... It marks a leaps forward
```

I claim the version before the change is a 'run-on sentence', i.e.
two separate sentences that are connected improperly (see
https://academicguides.waldenu.edu/writingcenter/grammar/runonsentences).

This commit changes how those two sentences are connected; in
particular, it separates the sentences with a period and provides a
subject for the second sentence.