Reformat code blocks. Show code on site rather than format text.
On page https://www.selenium.dev/documentation/webdriver/drivers/
in section 'Quitting Sessions', the Java code example shows text:

```
{< gh-codeblock path="examples/java/src/test/java/dev/selenium/getting_started/FirstScript.java#L29" >}
```

Instead of this, we want to show the code from that file.
This commit adds an extra layer of `{}` in order to show the code.
The panels for other languages that work (i.e. Python, Ruby) have
this extra layer of `{}`.